requirements.txt
torch
numpy
python-chess
pybind11
